// src/app/models/flight/flightList.ts
export class FlightList {
  id: number;
  airlineLogo: string;
  airlineName: string;
  departureCityAirport: string;
  departureTime: string;
  arrivalTime: string;
  arrivalCityAirport: string;
  stops: number;
  stopCity: string[];
  stopDuration: number;
  departureCity: string;
  arrivalCity: string;
  price: string;
  duration: string;
}